requirements.txt
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==8.2.0
Pillow==9.5.0
